- MGR: Fix a crash in the simple GUI.         NOTE: Do not use the at() function on a vector unless you         also use a try/catch handler.     - MGR: Remove code duplication
    clientgui/
        sg_ClientStateIndicator.h
        sg_ProgressBar.cpp, .h
        sg_ViewTabPage.cpp

svn path=/trunk/boinc/; revision=14936

+++ clientgui/sg_ClientStateIndicator.h
@@ -32,7 +32,7 @@ class ClientStateIndicator : public wxPanel{
 public: 
 		int connIndicatorWidth;
 		int connIndicatorHeight;
-		double numOfIndic;
+		int numOfIndic;
 		int rightPosition;
 	    int topPosition;
 		int indexIndVis;
